Adding license file to lr archive.zip on publishing. #75
