+++ metashare/static/metashare/licences/publicDomain.txt
@@ -1,0 +1,2 @@
+Licence Agreement – Public Domain 
+The resource is free of all known legal restrictions.
